Maint/condense rcons (#290)
Merge extended/recorded RCON classes.
