+++ rcon/auto_kick.py
@@ -6,13 +6,13 @@ from rcon.discord import send_to_discord_audit
 from rcon.game_logs import on_connected
 from rcon.hooks import inject_player_ids
 from rcon.player_history import get_player_profile, player_has_flag
-from rcon.recorded_commands import RecordedRcon
+from rcon.rcon import Rcon
 from rcon.settings import SERVER_INFO
 
 logger = logging.getLogger(__name__)
 
 
-recorded_rcon = RecordedRcon(SERVER_INFO)
+recorded_rcon = Rcon(SERVER_INFO)
 
 
 @on_connected
