+++ rcon/cli.py
@@ -12,7 +12,7 @@ from rcon.automods import automod
 from rcon.cache_utils import RedisCached, get_redis_pool
 from rcon.game_logs import LogLoop
 from rcon.models import install_unaccent
-from rcon.recorded_commands import RecordedRcon
+from rcon.rcon import Rcon
 from rcon.scoreboard import live_stats_loop
 from rcon.server_stats import (
     save_server_stats_for_last_hours,
@@ -31,7 +31,7 @@ def cli():
     pass
 
 
-ctl = RecordedRcon(SERVER_INFO)
+ctl = Rcon(SERVER_INFO)
 
 
 @cli.command(name="live_stats_loop")
@@ -232,8 +232,8 @@ for name in dir(ctl):
     func = getattr(ctl, name)
 
     if (
-            not any(name.startswith(prefix) for prefix in PREFIXES_TO_EXPOSE)
-            or name in EXCLUDED
+        not any(name.startswith(prefix) for prefix in PREFIXES_TO_EXPOSE)
+        or name in EXCLUDED
     ):
         continue
     wrapped = do_print(func)
