+++ rcon/rcon.py
@@ -1,22 +1,40 @@
 import logging
 import os
 import re
+from concurrent.futures import ThreadPoolExecutor, as_completed
 from datetime import datetime, timedelta
+from functools import cached_property
 from time import sleep
-from typing import Dict, Iterable, List, Optional, Tuple, Union, TypedDict
+from typing import Iterable
+
+from dateutil import parser, relativedelta
 
 from rcon.cache_utils import get_redis_client, invalidates, ttl_cache
 from rcon.commands import CommandFailedError, ServerCtl, VipId
-from rcon.models import PlayerVIP, enter_session
-from rcon.steam_utils import get_player_country_code, get_player_has_bans
+from rcon.config import get_config
+from rcon.models import AdvancedConfigOptions, PlayerSteamID, PlayerVIP, enter_session
+from rcon.player_history import (
+    add_player_to_blacklist,
+    get_profiles,
+    safe_save_player_action,
+)
+from rcon.steam_utils import (
+    get_player_country_code,
+    get_player_has_bans,
+    get_players_country_code,
+    get_players_have_bans,
+)
 from rcon.types import (
+    EnrichedGetPlayersType,
     GameState,
+    GetDetailedPlayer,
+    GetDetailedPlayers,
+    GetPlayersType,
     ParsedLogsType,
     StructuredLogLineType,
     StructuredLogLineWithMetaData,
-    GetDetailedPlayer,
 )
-from rcon.utils import get_server_number
+from rcon.utils import ALL_ROLES, ALL_ROLES_KEY_INDEX_MAP, get_server_number
 
 STEAMID = "steam_id_64"
 NAME = "name"
@@ -97,8 +115,199 @@ class Rcon(ServerCtl):
         r"MESSAGE: player \[(.+)\((\d+)\)\], content \[(.+)\]", re.DOTALL
     )
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, *args, pool_size: bool | None = None, **kwargs):
         super().__init__(*args, **kwargs)
+
+        # config/default_config.yml config.yml, etc.
+        config = get_config()
+        try:
+            self.advanced_settings = AdvancedConfigOptions(
+                **config["ADVANCED_CRCON_SETTINGS"]
+            )
+        except ValueError as e:
+            # This might look dumb but pydantic provides useful error messages in the
+            # stack trace and we don't have to remember to keep updating this if we add
+            # any more fields to the ADVANCED_CRCON_SETTINGS config
+            logger.exception(e)
+
+        if pool_size is not None:
+            self.pool_size = pool_size
+        else:
+            self.pool_size = self.advanced_settings.thread_pool_size
+
+    @cached_property
+    def thread_pool(self):
+        return ThreadPoolExecutor(self.pool_size)
+
+    def run_in_pool(self, function_name: str, *args, **kwargs):
+        return self.thread_pool.submit(getattr(self, function_name), *args, **kwargs)
+
+    def get_players_fast(self) -> list[GetPlayersType]:
+        players = {}
+        ids = []
+
+        for name, steam_id_64 in self.get_playerids():
+            players[steam_id_64] = {NAME: name, STEAMID: steam_id_64}
+            ids.append(steam_id_64)
+
+        countries = self.thread_pool.submit(get_players_country_code, ids)
+        bans = self.thread_pool.submit(get_players_have_bans, ids)
+
+        for future in as_completed([countries, bans]):
+            d = future.result()
+            for steamid, data in d.items():
+                players.get(steamid, {}).update(data)
+
+        return list(players.values())
+
+    @ttl_cache(ttl=5)
+    def get_players(self) -> list[EnrichedGetPlayersType]:
+        players = self.get_players_fast()
+
+        vips = set(v["steam_id_64"] for v in super().get_vip_ids())
+        steam_ids = [p.get(STEAMID) for p in players if p.get(STEAMID)]
+        profiles = {p["steam_id_64"]: p for p in get_profiles(steam_ids)}
+
+        updated_players: list[EnrichedGetPlayersType] = []
+
+        for p in players:
+            try:
+                updated_players.append(
+                    {
+                        **p,
+                        "profile": profiles[p["steam_id_64"]],
+                        "is_vip": p.get(STEAMID) in vips,
+                    }
+                )
+            except KeyError:
+                logger.error(f"Unable to retrieve profile information for player {p}")
+
+        return updated_players
+
+    def get_detailed_players(self) -> GetDetailedPlayers:
+        players = self.get_players_fast()
+        fail_count = 0
+        players_by_id: dict[str, GetDetailedPlayer] = {}
+
+        futures = {
+            self.run_in_pool("get_detailed_player_info", player[NAME]): player
+            for _, player in enumerate(players)
+        }
+
+        for future in as_completed(futures):
+            try:
+                player_data: GetDetailedPlayer = future.result()
+            except Exception:
+                logger.exception("Failed to get info for %s", futures[future])
+                fail_count += 1
+                player_data = self._get_default_info_dict(futures[future][NAME])
+            player = futures[future]
+            player.update(player_data)
+            players_by_id[player[STEAMID]] = player
+
+        return {
+            "players": players_by_id,
+            "fail_count": fail_count,
+        }
+
+    @ttl_cache(ttl=2, cache_falsy=False)
+    def get_team_view(self):
+        teams = {}
+        detailed_players = self.get_detailed_players()
+        players_by_id = detailed_players["players"]
+        fail_count = detailed_players["fail_count"]
+
+        logger.debug("Getting DB profiles")
+        steam_profiles = {
+            profile[STEAMID]: profile
+            for profile in get_profiles(list(players_by_id.keys()))
+        }
+
+        logger.debug("Getting VIP list")
+        try:
+            vips = set(v[STEAMID] for v in super().get_vip_ids())
+        except Exception:
+            logger.exception("Failed to get VIPs")
+            vips = set()
+
+        for player in players_by_id.values():
+            steam_id_64 = player[STEAMID]
+            profile = steam_profiles.get(player.get("steam_id_64"), {}) or {}
+            enriched_player: EnrichedGetPlayersType = {
+                "name": "",
+                "steam_id_64": steam_id_64,
+                "country": "",
+                "steam_bans": None,
+                "profile": profile,
+                "is_vip": steam_id_64 in vips,
+            }
+
+            teams.setdefault(player.get("team"), {}).setdefault(
+                player.get("unit_name"), {}
+            ).setdefault("players", []).append(player)
+
+        for team, squads in teams.items():
+            if team is None:
+                continue
+            for squad_name, squad in squads.items():
+                squad["players"] = sorted(
+                    squad["players"],
+                    key=lambda player: (
+                        ALL_ROLES_KEY_INDEX_MAP.get(player.get("role"), len(ALL_ROLES)),
+                        player.get("steam_id_64"),
+                    ),
+                )
+                squad["type"] = self._guess_squad_type(squad)
+                squad["has_leader"] = self._has_leader(squad)
+
+                try:
+                    squad["combat"] = sum(p["combat"] for p in squad["players"])
+                    squad["offense"] = sum(p["offense"] for p in squad["players"])
+                    squad["defense"] = sum(p["defense"] for p in squad["players"])
+                    squad["support"] = sum(p["support"] for p in squad["players"])
+                    squad["kills"] = sum(p["kills"] for p in squad["players"])
+                    squad["deaths"] = sum(p["deaths"] for p in squad["players"])
+                except Exception as e:
+                    logger.exception(e)
+
+        game = {}
+        for team, squads in teams.items():
+            if team is None:
+                continue
+            commander = [
+                squad for _, squad in squads.items() if squad["type"] == "commander"
+            ]
+            if not commander:
+                commander = None
+            else:
+                commander = (
+                    commander[0]["players"][0] if commander[0].get("players") else None
+                )
+
+            game[team] = {
+                "squads": {
+                    squad_name: squad
+                    for squad_name, squad in squads.items()
+                    if squad["type"] != "commander"
+                },
+                "commander": commander,
+                "combat": sum(s["combat"] for s in squads.values()),
+                "offense": sum(s["offense"] for s in squads.values()),
+                "defense": sum(s["defense"] for s in squads.values()),
+                "support": sum(s["support"] for s in squads.values()),
+                "kills": sum(s["kills"] for s in squads.values()),
+                "deaths": sum(s["deaths"] for s in squads.values()),
+                "count": sum(len(s["players"]) for s in squads.values()),
+            }
+
+        return dict(fail_count=fail_count, **game)
+
+    @ttl_cache(ttl=2)
+    def get_structured_logs(
+        self, since_min_ago, filter_action=None, filter_player=None
+    ) -> ParsedLogsType:
+        raw = super().get_logs(since_min_ago)
+        return self.parse_logs(raw, filter_action, filter_player)
 
     def get_admin_groups(self):
         # Defined here to avoid circular imports with commands.py
@@ -402,14 +611,14 @@ class Rcon(ServerCtl):
         bans.reverse()
         return temp_bans + bans
 
-    def do_unban(self, steam_id_64) -> List[str]:
+    def do_unban(self, steam_id_64) -> list[str]:
         """Remove all temporary and permanent bans from the steam_id_64"""
         bans = self.get_bans()
         type_to_func = {
             "temp": self.do_remove_temp_ban,
             "perma": self.do_remove_perma_ban,
         }
-        failed_ban_removals: List[str] = []
+        failed_ban_removals: list[str] = []
         for b in bans:
             if b.get("steam_id_64") == steam_id_64:
                 # TODO: This is no longer true as of U13
@@ -437,11 +646,11 @@ class Rcon(ServerCtl):
         return list(filter(lambda x: x.get("steam_id_64") == steam_id_64, bans))
 
     @ttl_cache(ttl=60 * 60)
-    def get_vip_ids(self) -> List[Dict[str, Union[str, Optional[datetime]]]]:
-        res: List[VipId] = super().get_vip_ids()
+    def get_vip_ids(self) -> list[dict[str, str | datetime | None]]:
+        res: list[VipId] = super().get_vip_ids()
         player_dicts = []
 
-        vip_expirations: Dict[str, datetime]
+        vip_expirations: dict[str, datetime]
         with enter_session() as session:
             server_number = get_server_number()
 
@@ -462,12 +671,116 @@ class Rcon(ServerCtl):
         return sorted(player_dicts, key=lambda d: d[NAME])
 
     def do_remove_vip(self, steam_id_64):
-        with invalidates(Rcon.get_vip_ids):
-            return super().do_remove_vip(steam_id_64)
+        """Removes VIP status on the game server and removes their PlayerVIP record."""
 
-    def do_add_vip(self, name, steam_id_64):
+        # Remove VIP before anything else in case we have errors
         with invalidates(Rcon.get_vip_ids):
-            return super().do_add_vip(steam_id_64, name)
+            result = super().do_remove_vip(steam_id_64)
+
+        server_number = get_server_number()
+        with enter_session() as session:
+            player: PlayerSteamID = (
+                session.query(PlayerSteamID)
+                .filter(PlayerSteamID.steam_id_64 == steam_id_64)
+                .one_or_none()
+            )
+            if player and player.vip:
+                logger.info(
+                    f"Removed VIP from {steam_id_64} expired: {player.vip.expiration}"
+                )
+                # TODO: This is an incredibly dumb fix because I can't get
+                # the changes to persist otherwise
+                vip_record: PlayerVIP = (
+                    session.query(PlayerVIP)
+                    .filter(
+                        PlayerVIP.playersteamid_id == player.id,
+                        PlayerVIP.server_number == server_number,
+                    )
+                    .one_or_none()
+                )
+                session.delete(vip_record)
+            elif player and not player.vip:
+                logger.warning(f"{steam_id_64} has no PlayerVIP record")
+            else:
+                # This is okay since you can give VIP to someone who has never been on a game server
+                # or that your instance of CRCON hasn't seen before, but you might want to prune these
+                logger.warning(f"{steam_id_64} has no PlayerSteamID record")
+
+        return result
+
+    def do_add_vip(self, name, steam_id_64, expiration: str = ""):
+        """Adds VIP status on the game server and adds or updates their PlayerVIP record."""
+        with invalidates(Rcon.get_vip_ids):
+            # Add VIP before anything else in case we have errors
+            result = super().do_add_vip(steam_id_64, name)
+
+        # postgres and Python have different max date limits
+        # https://docs.python.org/3.8/library/datetime.html#datetime.MAXYEAR
+        # https://www.postgresql.org/docs/12/datatype-datetime.html
+
+        server_number = get_server_number()
+        # If we're unable to parse the date, treat them as indefinite VIPs
+        expiration_date: str | datetime
+        try:
+            expiration_date = parser.parse(expiration)
+        except (parser.ParserError, OverflowError):
+            logger.warning(f"Unable to parse {expiration=} for {name=} {steam_id_64=}")
+            # For our purposes (human lifespans) we can use 200 years in the future as
+            # the equivalent of indefinite VIP access
+            expiration_date = datetime.utcnow() + relativedelta.relativedelta(years=200)
+
+        # Find a player and update their expiration date if it exists or create a new record if not
+        with enter_session() as session:
+            player: PlayerSteamID = (
+                session.query(PlayerSteamID)
+                .filter(PlayerSteamID.steam_id_64 == steam_id_64)
+                .one_or_none()
+            )
+
+            # If there's no PlayerSteamID record the player has never been on the server before
+            # Adding VIP to them is still valid as it's tracked on the game server
+            # The service that prunes expired VIPs checks for VIPs without a PlayerVIP record
+            # and creates them there as indefinite VIPs
+            if player:
+                vip_record: PlayerVIP = (
+                    session.query(PlayerVIP)
+                    .filter(
+                        PlayerVIP.server_number == server_number,
+                        PlayerVIP.playersteamid_id == player.id,
+                    )
+                    .one_or_none()
+                )
+                if not vip_record:
+                    new_vip_record = PlayerVIP(
+                        expiration=expiration_date,
+                        playersteamid_id=player.id,
+                        server_number=server_number,
+                    )
+                    logger.info(
+                        f"Added new PlayerVIP record {player.steam_id_64=} {expiration_date=}"
+                    )
+                    # TODO: This is an incredibly dumb fix because I can't get
+                    # the changes to persist otherwise
+                    session.add(new_vip_record)
+                else:
+                    previous_expiration = vip_record.expiration.isoformat()
+
+                    new_vip_record = PlayerVIP(
+                        expiration=expiration_date,
+                        playersteamid_id=player.id,
+                        server_number=server_number,
+                    )
+
+                    # TODO: This is an incredibly dumb fix because I can't get
+                    # the changes to persist otherwise
+                    session.delete(vip_record)
+                    session.commit()
+                    session.add(new_vip_record)
+                    logger.info(
+                        f"Modified PlayerVIP record {player.steam_id_64=} {player.vip.expiration=} {previous_expiration=}"
+                    )
+
+        return result
 
     def do_remove_all_vips(self):
         vips = self.get_vip_ids()
@@ -478,6 +791,26 @@ class Rcon(ServerCtl):
                 raise
 
         return "SUCCESS"
+
+    def do_message_player(
+        self,
+        player=None,
+        steam_id_64=None,
+        message: str = "",
+        by: str = "",
+        save_message: bool = False,
+    ):
+        res = super().do_message_player(player, steam_id_64, message)
+        if save_message:
+            safe_save_player_action(
+                rcon=self,
+                player_name=player,
+                action_type="MESSAGE",
+                reason=message,
+                by=by,
+                steam_id_64=steam_id_64,
+            )
+        return res
 
     def get_gamestate(self) -> GameState:
         """
@@ -529,14 +862,14 @@ class Rcon(ServerCtl):
         }
 
     @ttl_cache(ttl=2, cache_falsy=False)
-    def team_sizes(self) -> Tuple[int, int]:
+    def team_sizes(self) -> tuple[int, int]:
         """Returns the number of allied/axis players respectively"""
         result = self.get_gamestate()
 
         return result["num_allied_players"], result["num_axis_players"]
 
     @ttl_cache(ttl=2, cache_falsy=False)
-    def get_team_objective_scores(self) -> Tuple[int, int]:
+    def get_team_objective_scores(self) -> tuple[int, int]:
         """Returns the number of objectives held by the allied/axis team respectively"""
         result = self.get_gamestate()
 
@@ -579,7 +912,7 @@ class Rcon(ServerCtl):
                 self.do_add_map_to_rotation(
                     map_name, maps[len(maps) - 1], maps.count(maps[len(maps) - 1])
                 )
-                if super().set_map(map_name) != "SUCCESS":
+                if res := super().set_map(map_name) != "SUCCESS":
                     raise CommandFailedError(res)
 
     @ttl_cache(ttl=10)
@@ -846,33 +1179,48 @@ class Rcon(ServerCtl):
         with invalidates(self.get_profanities):
             return super().do_ban_profanities(",".join(profanities))
 
-    def do_punish(self, player, reason):
-        return super().do_punish(player, reason)
+    def do_punish(self, player, reason, by):
+        res = super().do_punish(player, reason)
+        safe_save_player_action(
+            rcon=self, player_name=player, action_type="PUNISH", reason=reason, by=by
+        )
+        return res
 
-    def do_switch_player_now(self, player):
+    def do_switch_player_now(self, player, by):
         return super().do_switch_player_now(player)
 
-    def do_switch_player_on_death(self, player):
+    def do_switch_player_on_death(self, player, by):
         return super().do_switch_player_on_death(player)
 
-    def do_kick(self, player, reason):
+    def do_kick(self, player, reason, by):
         with invalidates(Rcon.get_players):
-            return super().do_kick(player, reason)
+            res = super().do_kick(player, reason)
+        safe_save_player_action(
+            rcon=self, player_name=player, action_type="KICK", reason=reason, by=by
+        )
+        return res
 
     def do_temp_ban(
-        self, player=None, steam_id_64=None, duration_hours=2, reason="", admin_name=""
+        self, player=None, steam_id_64=None, duration_hours=2, reason="", by=""
     ):
         with invalidates(Rcon.get_players, Rcon.get_temp_bans):
             if player and re.match(r"\d+", player):
                 info = self.get_player_info(player)
                 steam_id_64 = info.get(STEAMID, None)
-                return super().do_temp_ban(
-                    None, steam_id_64, duration_hours, reason, admin_name
-                )
 
-            return super().do_temp_ban(
-                player, steam_id_64, duration_hours, reason, admin_name
+            res = super().do_temp_ban(
+                player, steam_id_64, duration_hours, reason, admin_name=by
             )
+
+            safe_save_player_action(
+                rcon=self,
+                player_name=player,
+                action_type="TEMPBAN",
+                reason=reason,
+                by=by,
+                steam_id_64=steam_id_64,
+            )
+            return res
 
     def do_remove_temp_ban(self, ban_log):
         with invalidates(Rcon.get_temp_bans):
@@ -882,14 +1230,32 @@ class Rcon(ServerCtl):
         with invalidates(Rcon.get_perma_bans):
             return super().do_remove_perma_ban(ban_log)
 
-    def do_perma_ban(self, player=None, steam_id_64=None, reason="", admin_name=""):
+    def do_perma_ban(self, player=None, steam_id_64=None, reason="", by=""):
         with invalidates(Rcon.get_players, Rcon.get_perma_bans):
             if player and re.match(r"\d+", player):
                 info = self.get_player_info(player)
                 steam_id_64 = info.get(STEAMID, None)
-                return super().do_perma_ban(None, steam_id_64, reason, admin_name)
 
-            return super().do_perma_ban(player, steam_id_64, reason, admin_name)
+            res = super().do_perma_ban(player, steam_id_64, reason, admin_name=by)
+            safe_save_player_action(
+                rcon=self,
+                player_name=player,
+                action_type="PERMABAN",
+                reason=reason,
+                by=by,
+                steam_id_64=steam_id_64,
+            )
+            try:
+                # TODO: this will never work when banning by name because they're already removed
+                # from the server before you can get their steam ID
+                if not steam_id_64:
+                    info = self.get_player_info(player)
+                    steam_id_64 = info["steam_id_64"]
+                # TODO add author
+                add_player_to_blacklist(steam_id_64, reason, by=by)
+            except:
+                logger.exception("Unable to blacklist")
+            return res
 
     @ttl_cache(60 * 5)
     def get_map_rotation(self):
@@ -1125,7 +1491,7 @@ class Rcon(ServerCtl):
         """Parse a chunk of raw gameserver RCON logs"""
         synthetic_actions = LOG_ACTIONS
         now = datetime.now()
-        parsed_log_lines: List[StructuredLogLineWithMetaData] = []
+        parsed_log_lines: list[StructuredLogLineWithMetaData] = []
         actions: set[str] = set()
         players: set[str] = set()
