+++ rconweb/rconweb/settings.py
@@ -21,7 +21,7 @@ from sentry_sdk import configure_scope
 from sentry_sdk.integrations.django import DjangoIntegration
 from sentry_sdk.integrations.logging import LoggingIntegration
 
-from rcon.extended_commands import Rcon
+from rcon.rcon import Rcon
 from rcon.settings import SERVER_INFO
 
 try:
